requirements.txt
fastapi==0.110.2
openai==1.23.4
pydantic==2.7.1
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.29
uvicorn==0.29.0
sentry-sdk==2.0.1
pytz==2024.1